implements prelaunch hook to comply with IHostAddon, deletes launcher action
