+++ client/ayon_comfyui/addon.py
@@ -1,16 +1,16 @@
 from pathlib import Path
-from ayon_core.addon import AYONAddon, IPluginPaths, IHostAddon
+from ayon_core.addon import AYONAddon, IHostAddon
 
 from .version import __version__
 
 
-ADDON_ROOT = Path(__file__).parent.resolve()
+ADDON_ROOT = Path(__file__).parent
 ADDON_NAME = "comfyui"
 ADDON_LABEL = "ComfyUI"
 ADDON_VERSION = __version__
 
 
-class ComfyUIAddon(AYONAddon, IPluginPaths, IHostAddon):
+class ComfyUIAddon(AYONAddon, IHostAddon):
     name = host_name = ADDON_NAME
     label = ADDON_LABEL
     version = __version__
@@ -20,14 +20,9 @@ class ComfyUIAddon(AYONAddon, IPluginPaths, IHostAddon):
         self.enabled = True
 
     def get_launch_hook_paths(self, app):
-        self.log.debug(f"{app = }")
         return [
             (ADDON_ROOT / "hooks").as_posix(),
         ]
 
-    def get_plugin_paths(self):
-        return {
-            "actions": [
-                (ADDON_ROOT / "plugins" / "actions").as_posix(),
-            ],
-        }
+    def get_workfile_extension(self) -> None:
+        return [".ps1"]
